Refactor JsonProcessor
